2.3.4 (12)

+++ pilot/eventservice/workexecutor/plugins/genericexecutor.py
@@ -330,7 +330,7 @@ class GenericExecutor(BaseExecutor):
 
             exit_code = None
             try:
-                iteration = 0L  # Python 2
+                iteration = long(0)  # Python 2
             except Exception:
                 iteration = 0  # Python 3
             while proc.is_alive():
